requirements.txt
discord.py==2.3.2
python-dotenv==1.0.0
aiohttp==3.9.3
fuzzywuzzy==0.18.0
python-Levenshtein==0.23.0 